move deep clustering model to contrib
